test: split api tests
